sbom-explorer: match exact sbom name (#23)
It is possible that there are additional SBOMs already uploaded in the
trustify instance used for test, and their names could be similar to the
one uploaded for testing, use exact matching when locating the item in
search results.

Othewise test can fail due to locating multiple matching candidates.

For example for sbomName='quarkus-bom' the test fails due to matching
results like:
- quarkus-bom
- quarkus-bom-2.13.8.Final-redhat-00004
- quarkus-bom-2.13.8.SP2-redhat-00001
- ... and many more like that